Move SanitizedParams class into utils
Better fit there as it's not really a 'model' so much as a bundle of
sanity-checking functions inside a class shell for convenient tracking
of vars.

Realistically, we only need the parse functions as so many endpoints
only use a few of the params.
